test: move resolve tests
