+++ test/core/src/external/pkg-browser/index.d.ts
