+++ test/core/src/external/pkg-browser/package.json
