+++ test/core/src/external/pkg-node/index.d.ts
