+++ test/core/src/external/pkg-node/package.json
