+++ test/core/test/resolve-ssr.test.ts
@@ -1,6 +1,8 @@
+// @vitest-environment node
+
 import { expect, test } from 'vitest'
-import pkgNode from '../pkg-node'
-import pkgBrowser from '../pkg-browser'
+import pkgNode from '../src/external/pkg-node'
+import pkgBrowser from '../src/external/pkg-browser'
 
 test('[ssr] resolves to ssr, when node is first in conditions', () => {
   expect(pkgNode).toBe('ssr')
