+++ test/core/test/resolve-web.test.ts
@@ -1,6 +1,8 @@
+// @vitest-environment happy-dom
+
 import { expect, test } from 'vitest'
-import pkgNode from '../pkg-node'
-import pkgBrowser from '../pkg-browser'
+import pkgNode from '../src/external/pkg-node'
+import pkgBrowser from '../src/external/pkg-browser'
 
 test('[web] resolves to ssr, when node is first in conditions', () => {
   expect(pkgNode).toBe('ssr')
