--- test/resolve/package.json
@@ -1,13 +1,0 @@
-{
-  "name": "@vitest/test-resolve",
-  "type": "module",
-  "private": true,
-  "scripts": {
-    "test": "vitest",
-    "coverage": "vitest run --coverage"
-  },
-  "devDependencies": {
-    "happy-dom": "^12.10.3",
-    "vitest": "workspace:*"
-  }
-}
