--- test/resolve/vitest.config.ts
@@ -1,15 +1,0 @@
-import { defineConfig } from 'vite'
-
-export default defineConfig({
-  test: {
-    environmentMatchGlobs: [
-      ['**/web.test.ts', 'happy-dom'],
-      ['**/ssr.test.ts', 'node'],
-    ],
-    server: {
-      deps: {
-        external: [/pkg-/],
-      },
-    },
-  },
-})
